Web-Crawler
A Web crawler, sometimes called a spider or spiderbot and often shortened to crawler, is an Internet bot that systematically browses the World Wide Web, typically operated by search engines for the purpose of Web indexing (web spidering). In this project, using Python, a Web-Crawler is created that fetches all the links from a particular website. (Use VS Code or PyCharm for getting the all the links from a website).